[Explorer] address search bar feedback
