+++ src/pages/layout/Search/SearchResultNotFound.tsx
@@ -1,0 +1,16 @@
+import {Box} from "@mui/material";
+import React from "react";
+
+export default function SearchResultNotFound() {
+  return (
+    <Box
+      sx={{
+        padding: 0.5,
+        display: "block",
+        width: "100%",
+      }}
+    >
+      No Results
+    </Box>
+  );
+}
